MemoryCacheLayer to use ConcurrentDictionary
Thanks to some super helpful insights from Nick Craver, it showed performance issues with the locking mechanism at a larger scale.

Switching back to ConcurrentDictionary shows Get being just over twice as fast in the CacheLayerComparison benchmark. Setting values is slightly faster. Evicting values is about the same.

Cleanup however takes a pretty large backwards step in both raw performance and memory usage. Likely the internal locking mechanisms are slower/more complicated for removes.

This shouldn't really be much of a problem - cleanup is an operation that happens in the background and the better locking mechanism of ConcurrentDictionary would still be a net positive.

In the alternative comparison benchmarks, it is about a 12% performance improvement.